moved files to src folder
added css and index.html
