+++ src/default.css
@@ -1,0 +1,3 @@
+body {
+    background-color: rgb(90, 79, 79);
+}
