+++ src/index.html
@@ -1,0 +1,10 @@
+<!DOCTYPE html>
+<head>
+    <link rel="stylesheet" href="/default.css">
+</head>
+<body>
+    <h1>Welcome to the new magic mirror</h1>
+    <div class="clock">
+        time
+    </div>
+</body>
